Move all output to a separate folder (/out by default)
With the move to notebooks / containers I banished the output to a
special folder (by default /out).  This required changing a whole lot
of places.

To reduce maintenance cost, I created a utility function to compute
the path of a generated file.

Things are a bit muddled because I considered output files generated
by make and files generated by python.  Which sort of makes sense for
me but it may not make sense at PurvisLab.