Compile R models into python modules before loading
Tuyrns out an R model is not equivalent to a python module for modelr.
The later has a bunch of functions which I haven't implemented for the
former.  For now, always compile R models into python modules so they
can be modeled correctly.

This replicates a bit of code (most of r2py.py) and so is a hack.  The
real solution is to implement the missing functions for R model read
from rds files but Adriana is running into problems and this hack should
let her move forward.