Fix handling of cells where hpd_ref is zero and hpd_proj is non-zero
To project human population density I multiply GRUMPS with a growth
factor--the ratio between the raw data (whether SPS or HYDE) between the
reference year and the projected year (could be in the past).

But this causes problems when the reference year is zero and the
projected year is non-zero--which happens mostly when projecting past
the reference year.

I had tried a number of hack, none of which was right.  I finally bit
the bullet and did the "correct" thing; use the projected raw data for
cells that meet this condition.  This required using a python class for
the projection, similar in spirit to how WPP works.