Sort level 0 columns so rasters come first
When a simple expression gets evaluated first, there's not way to
determine what the output shape should be.  I had tried fixing this a
couple of days ago but that was a brain fart.

This change sorts level 0 columns so rasters always come first.  This
will guarantee the shape of the output is know before computing a siple
expression.