Use cartopy to warp the image
Cartopy has the ability to warp the image.  This removes a bit of code
and makes things simpler. But, it is a bit slower.  The projection seems
to run in python code which is slower than GDAL C code.  At least,
that's my thinking.

SO I am checking this in and will decide later whether to keep it or go
back to rasterio for warping.