Add support for 0.5° resolution and cleanup
The output for 0.5° resolution had unaligned pixels.  When I tried using
GDAL to do the downsampling it always produced artifacts and caused the
sum to be significantly different (the SPS dataset record persons not
persons / km^2).

So use instead a simple numpy reshape script to sum the cells.  This
requires that the scaling factor be an integer.  And, to simplify
things, I made the output cover the whole world.  It's a bit wasteful
with nodata but compression takes care of it.