Change time units and bounds
Time units of "years since" are non-standard and some
tools (e.g. rioxarray) refuse to parse the time dimension.  So change
the time units to be "days since" the start of epoch.  Because SPS is
only a future dataset I don't need to consider dates before 1970.

Also change the bounds of the data.  I originally expanded the dataset
to cover the whole world but I now find it more convenient to keep the
specific bounds for which there is data.  This requires possibly reading
a few extra rows of input when the number of rows is not a multiple of
the factor between resolutions.