Add a new class (LUI) which can be used to generate land-use intensity
This change adds a new class (LUI) which is a more generic way of
generating land-use intensity data.  Rather than hard-coding names to
models, this class takes as input the name of the variabe to use and the
name of the model.  When evaluated it runs the LUI model passing in the
specified variable as the land-use class.